Make webui account config more secure
Create the account.conf file only as a sample, so that users can change it
locally and do not have the risk of commiting it to the repo.

TODO: Add a step into the documentation to copy the sample file (and hopefully
change the PW to be a bit more secure).